refactor:  역할에 따라 뷰 분리
- 역할 단위 분리로 명확한 책임
- 파일 크기 줄이기
