+++ Week_07/Week-07-Info.plist
@@ -4,12 +4,9 @@
 <dict>
 	<key>UIAppFonts</key>
 	<array>
-		<string>Pretendard-Black</string>
 		<string>Pretendard-Bold</string>
-		<string>Pretendard-ExtraBold</string>
-		<string>Pretendard-ExtraLight</string>
-		<string>Pretendard-Light</string>
 		<string>Pretendard-Medium</string>
+		<string>Pretendard-Light</string>
 	</array>
 </dict>
 </plist>
